requirements.txt
django
psycopg2-binary
pillow
django-ninja
django-ninja-jwt
python-dotenv